Extract local date persistence value handlers (#105)
