+++ aes-core/src/main/java/com/ijioio/aes/core/persistence/jdbc/value/handler/JdbcBooleanPersistenceValueHandler.java
@@ -110,5 +110,4 @@ public class JdbcBooleanPersistenceValueHandler extends BaseJdbcPersistenceValue
 			throw new PersistenceException(e);
 		}
 	}
-
 }
