+++ aes-core/src/main/java/com/ijioio/aes/core/persistence/jdbc/value/handler/JdbcBytePersistenceValueHandler.java
@@ -110,5 +110,4 @@ public class JdbcBytePersistenceValueHandler extends BaseJdbcPersistenceValueHan
 			throw new PersistenceException(e);
 		}
 	}
-
 }
