+++ aes-core/src/main/java/com/ijioio/aes/core/persistence/jdbc/value/handler/JdbcCharacterPersistenceValueHandler.java
@@ -113,5 +113,4 @@ public class JdbcCharacterPersistenceValueHandler extends BaseJdbcPersistenceVal
 			throw new PersistenceException(e);
 		}
 	}
-
 }
