+++ aes-core/src/main/java/com/ijioio/aes/core/persistence/jdbc/value/handler/JdbcDoublePersistenceValueHandler.java
@@ -110,5 +110,4 @@ public class JdbcDoublePersistenceValueHandler extends BaseJdbcPersistenceValueH
 			throw new PersistenceException(e);
 		}
 	}
-
 }
