+++ aes-core/src/main/java/com/ijioio/aes/core/persistence/jdbc/value/handler/JdbcFloatPersistenceValueHandler.java
@@ -109,5 +109,4 @@ public class JdbcFloatPersistenceValueHandler extends BaseJdbcPersistenceValueHa
 			throw new PersistenceException(e);
 		}
 	}
-
 }
