+++ aes-core/src/main/java/com/ijioio/aes/core/persistence/jdbc/value/handler/JdbcIntegerPersistenceValueHandler.java
@@ -110,5 +110,4 @@ public class JdbcIntegerPersistenceValueHandler extends BaseJdbcPersistenceValue
 			throw new PersistenceException(e);
 		}
 	}
-
 }
