+++ aes-core/src/main/java/com/ijioio/aes/core/persistence/jdbc/value/handler/JdbcLongPersistenceValueHandler.java
@@ -110,5 +110,4 @@ public class JdbcLongPersistenceValueHandler extends BaseJdbcPersistenceValueHan
 			throw new PersistenceException(e);
 		}
 	}
-
 }
